+++ api/docs/visao_geral/chapters/14-sintese_e_direcao.md
@@ -1,0 +1,180 @@
+# Síntese Geral e Direção de Evolução
+
+Este capítulo finaliza a visão geral da API reunindo os principais pontos apresentados ao longo do documento. O objetivo é consolidar um entendimento comum sobre o estágio atual da aplicação e deixar explícita a direção que deve orientar sua evolução.
+
+## O que a API é hoje
+
+No estágio atual, a API do projeto `banklab` deve ser entendida como um backend funcional para um core bancário simplificado, e não apenas como um conjunto de endpoints independentes.
+
+Ela já possui:
+
+- autenticação com sessão e refresh token;
+- vínculo entre usuário autenticado e customer;
+- operações administrativas de aprovação;
+- criação e listagem de contas;
+- consulta de saldo;
+- depósito, saque e transferência;
+- extrato baseado em ledger;
+- proteção transacional para fluxos sensíveis;
+- separação arquitetural por módulos e camadas;
+- testes distribuídos por responsabilidade;
+- documentação técnica em evolução.
+
+Essa combinação indica que o projeto já exige disciplina arquitetural real. Mesmo sem cobrir um domínio bancário amplo, ele já lida com identidade, ownership, persistência crítica, consistência financeira e organização modular.
+
+## O que precisa ser preservado
+
+Ao evoluir a aplicação, alguns princípios devem ser preservados para manter a coerência do sistema.
+
+### 1. Separação por módulos
+
+Os módulos `auth`, `account`, `customer` e `admin` não são apenas agrupamentos de arquivo. Eles representam contextos de responsabilidade.
+
+Novas funcionalidades devem respeitar essa divisão sempre que possível, evitando concentrar no módulo errado regras que pertencem a outro domínio.
+
+### 2. Separação por camadas
+
+As camadas `delivery`, `application`, `domain` e `infrastructure` existem para separar:
+
+- adaptação HTTP;
+- coordenação de caso de uso;
+- regra de negócio;
+- implementação técnica.
+
+Esse desenho só funciona bem se novas mudanças mantiverem essa distinção. Handler não deve virar centro de regra. SQL não deve virar centro de política de negócio. O domínio não deve depender de detalhes técnicos.
+
+### 3. Contexto autenticado como base de identidade
+
+O sistema foi modelado para derivar identidade sensível a partir do contexto autenticado sempre que possível.
+
+Isso significa que novas features devem tratar com cuidado a necessidade de receber do cliente dados como `customer_id` quando o backend já conhece esse escopo a partir da sessão autenticada.
+
+### 4. Ownership como regra estrutural
+
+Autenticação não substitui autorização nem ownership.
+
+Se um fluxo opera sobre conta, customer ou recurso específico, a aplicação precisa continuar verificando a relação entre o usuário autenticado e o recurso acessado.
+
+### 5. Consistência financeira como requisito de arquitetura
+
+Saldo, ledger, transação, lock e idempotência não são detalhes opcionais. Eles fazem parte da forma como o sistema preserva integridade.
+
+Qualquer funcionalidade que altere estado financeiro precisa considerar explicitamente:
+
+- atomicidade;
+- concorrência;
+- coerência entre saldo atual e histórico;
+- possibilidade de repetição de requisição.
+
+## O que a arquitetura privilegia
+
+A arquitetura atual privilegia:
+
+- simplicidade operacional;
+- clareza de responsabilidade;
+- consistência transacional local;
+- testabilidade;
+- evolução incremental;
+- onboarding mais fácil.
+
+Essas qualidades fazem bastante sentido para o estágio atual do projeto. Elas permitem que a aplicação cresça com uma base compreensível e relativamente segura.
+
+## O que a arquitetura ainda não busca resolver plenamente
+
+Ao mesmo tempo, o sistema ainda não está orientado a resolver, como prioridade principal:
+
+- microsserviços distribuídos;
+- deploy independente por módulo;
+- observabilidade avançada;
+- processamento fortemente assíncrono;
+- event sourcing completo;
+- CQRS amplo;
+- múltiplos subdomínios financeiros complexos.
+
+Isso não representa deficiência arquitetural. Representa um recorte intencional.
+
+A aplicação está organizada para resolver primeiro o que já existe de forma consistente, sem antecipar complexidades que ainda não são exigidas pelo estágio atual do produto.
+
+## Como evoluir sem perder coerência
+
+Uma evolução saudável da API tende a seguir alguns passos de raciocínio.
+
+### Primeiro: identificar o dono da regra
+
+Antes de implementar uma mudança, vale perguntar:
+
+- essa regra pertence a `auth`, `account`, `customer` ou `admin`?
+- ela é realmente transversal ou apenas parece transversal?
+
+Essa pergunta ajuda a evitar que o código cresça em locais convenientes, mas incorretos.
+
+### Segundo: identificar o nível da mudança
+
+Também é útil perguntar:
+
+- isso é adaptação HTTP?
+- isso é caso de uso?
+- isso é regra de domínio?
+- isso é persistência?
+
+Essa pergunta ajuda a preservar a anatomia interna dos módulos.
+
+### Terceiro: identificar impacto sobre identidade e ownership
+
+Se a operação depende do usuário autenticado, de `role`, de `customer_id` ou de escopo do recurso, essa dimensão precisa ser tratada explicitamente no desenho da feature.
+
+### Quarto: identificar impacto sobre consistência
+
+Se a mudança toca saldo, conta, histórico financeiro ou múltiplas entidades persistidas, ela precisa ser avaliada também do ponto de vista transacional e não apenas funcional.
+
+### Quinto: identificar o nível correto de teste e documentação
+
+Toda mudança relevante deve considerar:
+
+- onde o comportamento precisa ser validado;
+- qual documento precisa ser atualizado para manter a visão geral e os materiais especializados coerentes.
+
+## Direção de evolução provável
+
+A partir da base atual, uma evolução natural da API pode incluir:
+
+- fortalecimento da observabilidade;
+- crescimento do módulo `account` com políticas mais ricas;
+- amadurecimento de monitoramento e métricas;
+- ampliação dos testes de integração em fluxos críticos;
+- evolução contínua da documentação;
+- refinamento das fronteiras entre módulos;
+- eventual avaliação de extrações mais fortes caso a pressão operacional do sistema aumente.
+
+O ponto importante é que essas evoluções não exigem ruptura com a arquitetura existente. Ao contrário, a estrutura atual foi pensada para permitir crescimento sem perder completamente o mapa conceitual do sistema.
+
+## O valor desta visão geral
+
+Este conjunto de capítulos foi construído para funcionar como um material de entendimento da aplicação no seu estado atual.
+
+Ele não substitui:
+
+- o código como fonte definitiva do comportamento;
+- a documentação REST detalhada;
+- a documentação de banco;
+- a análise direta dos testes e dos casos de uso.
+
+Mas ele reduz significativamente o custo de chegar até esse entendimento. Seu valor está em conectar:
+
+- arquitetura;
+- domínio;
+- organização do projeto;
+- fluxo de requisições;
+- persistência;
+- testes;
+- documentação.
+
+Essa conexão é o que permite a um novo desenvolvedor enxergar o sistema como unidade coerente, e não apenas como uma coleção de arquivos.
+
+## Síntese final
+
+A API do `banklab` é, hoje, um monolito modular em Go com separação clara entre camadas, foco em coerência de domínio e atenção explícita a identidade, ownership e consistência financeira.
+
+Seu tamanho ainda é controlado, mas sua estrutura já reflete preocupações que vão além de um CRUD simples. O sistema organiza autenticação, customer, conta e administração de forma relativamente coesa, e usa PostgreSQL, transações e ledger para sustentar operações sensíveis.
+
+Para novos desenvolvedores, a principal orientação é simples: antes de implementar uma mudança, entender o significado da regra é tão importante quanto entender a técnica usada para realizá-la. A arquitetura atual oferece um bom mapa para isso. Preservar esse mapa é parte do próprio trabalho de evolução da aplicação.
