+++ api/docs/visao_geral/chapters/16-conclusao.md
@@ -1,0 +1,27 @@
+# Conclusão
+
+Esta visão geral foi construída para funcionar como um mapa de entendimento da API no seu estágio atual. Ao longo dos capítulos, o objetivo foi sair de uma descrição superficial da aplicação e chegar a uma leitura mais estruturada sobre:
+
+- o que a API faz;
+- como ela está organizada;
+- quais decisões arquiteturais sustentam o sistema;
+- como os principais fluxos se comportam;
+- quais cuidados precisam ser preservados ao evoluir o código.
+
+O ponto central desta documentação é que a aplicação já não deve ser tratada como um conjunto casual de endpoints ou como um CRUD genérico. Mesmo com escopo ainda controlado, ela já possui:
+
+- separação por módulos de negócio;
+- separação por camadas internas;
+- autenticação e contexto de identidade;
+- regras de ownership;
+- operações financeiras com preocupação transacional;
+- modelagem explícita de saldo e ledger;
+- testes e documentação como parte da sustentação do sistema.
+
+Isso significa que contribuir com a API exige atenção não apenas à implementação técnica imediata, mas também à coerência arquitetural da mudança. Em muitos casos, entender **por que** uma responsabilidade está em determinado módulo ou camada é tão importante quanto entender **como** aquela parte foi codificada.
+
+Para novos desenvolvedores, a principal utilidade deste material é reduzir o custo de interpretação do projeto. Em vez de começar diretamente por arquivos isolados, esta visão geral oferece um modelo mental de referência para navegar no código com mais contexto, identificar melhor o lugar de cada regra e tomar decisões de implementação com mais segurança.
+
+Para a continuidade do projeto, o valor desta documentação está em preservar um entendimento compartilhado. À medida que a API evoluir, novos endpoints, novas regras e novas decisões de modelagem poderão ser incorporados sem perder o fio condutor da arquitetura, desde que essa coerência continue sendo registrada e mantida.
+
+Em síntese, esta visão geral não encerra o entendimento do sistema, mas estabelece uma base comum para estudá-lo, mantê-lo e evoluí-lo. O código continua sendo a fonte definitiva do comportamento da aplicação; esta documentação existe para tornar esse código mais legível do ponto de vista arquitetural, funcional e organizacional.
