+++ mermaid-images/31214cedf4439210275feab44f469949e4380659.mmd
@@ -1,0 +1,8 @@
+flowchart LR
+    Client[Mobile / Client<br>Flutter] -->|HTTP| API[Bank API<br>GoLang]
+    API -->|Transactions| DB[(Database<br>PostgreSQL)]
+
+    subgraph Banklab
+        API
+        DB
+    end
