+++ mermaid-images/4ece891a70a04a4c84165783352570f14c60f0e0.mmd
@@ -1,0 +1,3 @@
+flowchart TD
+    Delivery --> Application --> Domain
+    Infrastructure --> Domain
